pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eostrata"
version = "0.1.0"
description = "One tool to fetch, store, aggregate, and serve earth observation layers."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "tqdm>=4.66",
    "rasterio>=1.3",
    "shapely>=2.0",
    "xarray>=2024.2",
    "zarr>=2.17,<3",
    "numpy>=1.26",
    "pystac>=1.10",
    "pydantic-settings>=2.2",
    "typer>=0.12",
    "rich>=13",
    "titiler-xarray[minimal]>=0.24.2",
    "titiler-core>=0.24.2",
    "stac-fastapi-api>=6.2.1",
    "stac-fastapi-types>=6.2.1",
    "uvicorn[standard]>=0.42.0",
    "rioxarray>=0.19.0",
    "attrs>=26.1.0",
    "h5py>=3.16.0",
]

[project.optional-dependencies]
scheduler = [
    "apscheduler>=3.10",
    "pyyaml>=6.0",
]
cds = ["cdsapi>=0.7"]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=6",
    "httpx>=0.27",
    "ruff>=0.9",
]

[project.scripts]
eostrata = "eostrata.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["eostrata*"]

# ── Ruff ──────────────────────────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long — formatter handles this
    "B008",   # do not perform function calls in default arguments (FastAPI pattern)
    "SIM108", # ternary operator — sometimes less readable
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",  # assert is fine in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── Pytest ────────────────────────────────────────────────────────────────────

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["eostrata"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
    "ruff>=0.15.7",
]
